applications/backend/requirements.txt
fastapi
uvicorn[standard]
azure-cosmos
pydantic

